fix(platform): preserve dangling symlink removal
